refactor(backend): move multipart writer lease and completion boundary out of routes
Finish reducing routes/mediaAssets.ts to an HTTP adapter: the writer lease, exactly-once
resolution, abort and completion application boundaries move into mediaAssets/multipart/.
The file drops 1,971 to 550 lines and now declares only MediaAssetsRoutesOptions and
createMediaAssetsRoutes. Nothing under mediaAssets/ imports upward from routes/ any more.
